hermitd/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hermitd"
version = "0.1.0"
description = "llm-powered shell assistant daemon"
authors = [{name = "Max Yu"}]
readme = "README.md"
requires-python = ">=3.9"
license = {text = "Apache-2.0"}
keywords = [
    "llmsh",
    "llm",
    "shell",
    "bash",
    "copilot",
]
dependencies = [
    "jinja2",
    "pydantic",
    "pyyaml",
    "pyzmq",
    "openai",
    "anthropic",
]

classifiers = [
  # How mature is this project? Common values are
  #   3 - Alpha
  #   4 - Beta
  #   5 - Production/Stable
  "Development Status :: 3 - Alpha",

  # Indicate who your project is intended for
  "Intended Audience :: Developers",
  "Topic :: Software Development :: Build Tools",

  # Only linux is supported for llmsh+hermitd currently
  "Operating System :: POSIX :: Linux",

  # Pick your license as you wish (see also "license" above)
  "License :: OSI Approved :: Apache Software License",

  # Specify the Python versions you support here.
  "Programming Language :: Python :: 3",
]

[project.urls]
"Repository" = "https://github.com/maxyu1115/llmsh"

[project.optional-dependencies]
local = [
  "vllm>=0.5.3",
  "transformers",
]

[project.scripts]
hermitd = "hermitd:main"
hermitd-install = "hermitd:install_hermitd"

[tool.setuptools]
packages = ["hermitd"]
include-package-data = true
package-data = {"hermitd" = ["hermitd.service", "hermitd.conf.template"]}
